requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# cog\Moderation.py: 16,17
APScheduler == 3.6.3
art==4.6

# OpticAfrica.py: 6
# cog\Filter.py: 4
# cog\Help.py: 7
# cog\Moderation.py: 7
# cog\NormalCommands.py: 7
# cog\Welcome Helper.py: 7
# cog\sadcats.py: 6
# cog\uwu.py: 6
aiohttp == 3.6.2

# cog\utils\db.py: 40
asyncpg == 0.20.1

# OpticAfrica.py: 3,4
# cog\Filter.py: 6,7
# cog\Help.py: 3,4,6
# cog\Moderation.py: 5
# cog\NormalCommands.py: 3,4,6
# cog\Welcome Helper.py: 3,4,6,14
# cog\sadcats.py: 3,4
# cog\utils\checks.py: 1
# cog\utils\context.py: 1,3
# cog\utils\functions.py: 2,3
# cog\utils\paginator.py: 2,3
# cog\utils\time.py: 5
# cog\uwu.py: 3,4
discord.py == 1.3.2

# cog\utils\cache.py: 8
lru-dict == 1.1.6

# cog\utils\maps.py: 5
lxml == 4.5.0

# cog\utils\time.py: 2
parsedatetime == 2.5

# cog\Filter.py: 8
profanity == 1.1

# cog\utils\cache.py: 8
py-lru-cache == 0.1.4

# cog\utils\time.py: 3
python_dateutil == 2.8.1

# cog\Moderation.py: 19
requests == 2.23.0
